build: re-structure repo to be pip-installable
This commit re-structures the repository's build and environment to
work more like the other JupyterLite kernels, where you can just
install them like other JupyterLab extensions.

The build system now uses Hatch, but still pulls relevant info from
package.json to keep the metadata DRY. Because it now uses Hatch,
the setup.py's contents are replaced with a pyproject.toml, jupyter-
packaging is replaced by the hatch-jupyter-builder, and the manifest
is no loonger needed.

For convenvience, an environment.yml is included to setup a conda
environment for developers.

Finally, the README is updated with the simpler install process, and
re-structured slightly to be more "public-facing" rather than
primarily "contributor-facing".